+++ src/main/java/de/uni_jena/cs/fusion/abecto/metaentity/Omission.java
@@ -27,27 +27,35 @@ import de.uni_jena.cs.fusion.abecto.sparq.SparqlPattern;
 @SparqlNamespace(prefix = "rdf", namespace = "http://www.w3.org/1999/02/22-rdf-syntax-ns#")
 @SparqlNamespace(prefix = "abecto", namespace = "http://fusion.cs.uni-jena.de/ontology/abecto#")
 public class Omission {
-	@SparqlPattern(predicate = "rdf:type", object = "abecto:Measurement")
+	@SparqlPattern(predicate = "rdf:type", object = "abecto:Omission")
 	public final Resource id;
-	/** The category of the missing resource. */
+	/** The category of the missing omittedResource. */
 	@SparqlPattern(subject = "id", predicate = "abecto:categoryName")
 	public final String categoryName;
-	/** The ontology which is missing the resource. */
+	/** The ontology which is missing the omittedResource. */
 	@SparqlPattern(subject = "id", predicate = "abecto:ontology")
 	public final UUID ontology;
-	/** The resource which is missing. */
-	@SparqlPattern(subject = "id", predicate = "abecto:resource1")
-	public final Resource resource;
-	/** The ontology which contains the missing resource. */
-	@SparqlPattern(subject = "id", predicate = "abecto:ontology")
+	/** The omittedResource which is missing. */
+	@SparqlPattern(subject = "id", predicate = "abecto:omittedResource")
+	public final Resource omittedResource;
+	/** The ontology which contains the omitted resource. */
+	@SparqlPattern(subject = "id", predicate = "abecto:sourceOntology")
 	public final UUID source;
 
+	public Omission() {
+		this.id = null;
+		this.categoryName = null;
+		this.ontology = null;
+		this.omittedResource = null;
+		this.source = null;
+	}
+	
 	public Omission(@Member("id") Resource id, @Member("categoryName") String categoryName,
-			@Member("ontology") UUID ontology, @Member("resource") Resource resource, @Member("source") UUID source) {
+			@Member("ontology") UUID ontology, @Member("omittedResource") Resource resource, @Member("source") UUID source) {
 		this.id = id;
 		this.categoryName = categoryName;
 		this.ontology = ontology;
-		this.resource = resource;
+		this.omittedResource = resource;
 		this.source = source;
 	}
 
@@ -58,12 +66,12 @@ public class Omission {
 		}
 		Omission o = (Omission) other;
 		return Objects.equals(this.categoryName, o.categoryName) && Objects.equals(this.ontology, o.ontology)
-				&& Objects.equals(this.resource, o.resource) && Objects.equals(this.source, o.source);
+				&& Objects.equals(this.omittedResource, o.omittedResource) && Objects.equals(this.source, o.source);
 	}
 
 	@Override
 	public int hashCode() {
-		return this.categoryName.hashCode() + this.ontology.hashCode() + this.resource.hashCode()
+		return this.categoryName.hashCode() + this.ontology.hashCode() + this.omittedResource.hashCode()
 				+ this.source.hashCode();
 	}
 }